+++ app/controllers/spree/retail/shopify/hooks/orders_controller.rb
@@ -1,0 +1,13 @@
+module Spree
+  module Retail
+    module Shopify
+      module Hooks
+        class OrdersController < HooksController
+          def create
+            head :ok
+          end
+        end
+      end
+    end
+  end
+end
