+++ config/routes.rb
@@ -1,3 +1,9 @@
 Spree::Core::Engine.routes.draw do
-  # Add your extension routes here
+  namespace :retail do
+    namespace :shopify do
+      namespace :hooks do
+        resource :order, only: :create
+      end
+    end
+  end
 end
